Fixed run source file to support output streaming
