+++ src/main/kotlin/com/nanovms/ops/command/Exceptions.kt
@@ -1,0 +1,5 @@
+package com.nanovms.ops.command
+
+import com.intellij.openapi.vfs.VirtualFile
+
+class UnsupportedSourceType(file: VirtualFile) : Exception("Unsupported source type: ${file.extension}") {}
